perf: Waiting page에서 WaitingListContainer와 WaitingBottomBarContainer 컴포넌트 분리
- 다른 사용자 입장, 퇴장 시 불필요한 헤더의 리렌더링 없앰
